live view for biotope
