+++ apps/xim2_web/lib/xim2_web/components/layouts/root.html.heex
@@ -11,7 +11,7 @@
     <script defer phx-track-static type="text/javascript" src={~p"/assets/app.js"}>
     </script>
   </head>
-  <body class="bg-slate-900 antialiased">
+  <body class="bg-slate-900 text-sky-100 antialiased">
     <%= @inner_content %>
   </body>
 </html>
