+++ apps/xim2_web/lib/xim2_web/controllers/page_html/home.html.heex
@@ -11,6 +11,7 @@
   <p>
     <ul>
       <li><.link class="text-sky-200" navigate={~p"/grid"}>Grid</.link></li>
+      <li><.link class="text-sky-200" navigate={~p"/biotope"}>Biotope</.link></li>
       <li><.link class="text-sky-200" navigate={~p"/colors"}>Colors</.link></li>
     </ul>
   </p>
