+++ apps/xim2_web/lib/xim2_web/router.ex
@@ -20,6 +20,7 @@ defmodule Xim2Web.Router do
     get "/", PageController, :home
     get "/colors", PageController, :colors
     live "/grid", GridLive.Show, :show
+    live "/biotope", BiotopeLive.Index, :index
   end
 
   # Other scopes may use custom stacks.
